backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-decouple==2.1
django-environ==0.7.0
django-filter==21.1
django-import-export==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.1.1
et-xmlfile==1.1.0
Faker==8.14.1
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.4.3
mixer==7.1.2
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tablib==3.0.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
